Revert "clean up and string pull working"
This reverts commit 7dbb6426f85e8926b26c3b6208dbba8ec414dcc8.

+++ NavMesh/LineRenderer/NavigationMesh.h
@@ -30,7 +30,7 @@ public:
 
 	void AddPoint(Vec2 point) { mPoints.push_back(point); }
 	void AddPointList(std::vector<Vec2> pointList);
-	int GetNumberOfTriangles() { return (int)mTriangles.size(); }
+	int GetNumberOfTriangles() { return mTriangles.size(); }
 	Triangle* GetTriangleAtIndex(int index) { return mTriangles[index]; }
 };
